restructured test module into multiple modules
